Add main menu with options
Now the main menu is opened when the program is started, and the main menu is shown when a game ends. The user can start a new game, show the leaderboard, or quit the program from the main menu.

Also renamed the "Main loop.py" to "main_loop.py" to make it easier to import from it.